# [PERF] Labels: faster encoding for -tags dedupelabels