Change from regular varint to a non-standard version: we encode small numbers (up to 32767) in 2 bytes, because we expect most Prometheus to have more than 127 unique strings, and we don't encode numbers larger than 4 bytes because we don't expect more than 536,870,912 unique strings.

A bit of copy-paste inlining because I couldn't persuade the Go compiler to do it for me.

Benchmarks need to be tweaked because otherwise they are optimised unrealistically for a tiny tiny Prometheus.

```
goos: linux
goarch: amd64
pkg: github.com/prometheus/prometheus/model/labels
cpu: Intel(R) Core(TM) i7-14700K
                                                 │  before.txt  │             after.txt              │
                                                 │    sec/op    │   sec/op     vs base               │
Labels_Get/with_5_labels/first_label/get-28         4.988n ± 1%   4.083n ± 1%  -18.14% (p=0.002 n=6)
Labels_Get/with_5_labels/first_label/has-28         3.334n ± 0%   2.978n ± 0%  -10.71% (p=0.002 n=6)
Labels_Get/with_5_labels/middle_label/get-28        8.957n ± 1%   7.563n ± 0%  -15.57% (p=0.002 n=6)
Labels_Get/with_5_labels/middle_label/has-28        7.840n ± 1%   6.451n ± 1%  -17.72% (p=0.002 n=6)
Labels_Get/with_5_labels/last_label/get-28          13.13n ± 0%   10.96n ± 1%  -16.53% (p=0.002 n=6)
Labels_Get/with_5_labels/last_label/has-28         12.210n ± 1%   9.869n ± 1%  -19.17% (p=0.002 n=6)
Labels_Get/with_5_labels/not-found_label/get-28     6.857n ± 4%   5.684n ± 1%  -17.11% (p=0.002 n=6)
Labels_Get/with_5_labels/not-found_label/has-28     6.906n ± 0%   5.585n ± 1%  -19.12% (p=0.002 n=6)
Labels_Get/with_10_labels/first_label/get-28        4.923n ± 1%   4.091n ± 1%  -16.91% (p=0.002 n=6)
Labels_Get/with_10_labels/first_label/has-28        3.331n ± 0%   2.975n ± 1%  -10.67% (p=0.002 n=6)
Labels_Get/with_10_labels/middle_label/get-28       16.65n ± 1%   13.71n ± 1%  -17.65% (p=0.002 n=6)
Labels_Get/with_10_labels/middle_label/has-28       15.74n ± 1%   12.65n ± 1%  -19.66% (p=0.002 n=6)
Labels_Get/with_10_labels/last_label/get-28         25.06n ± 1%   20.70n ± 0%  -17.39% (p=0.002 n=6)
Labels_Get/with_10_labels/last_label/has-28         24.50n ± 0%   19.47n ± 0%  -20.55% (p=0.002 n=6)
Labels_Get/with_10_labels/not-found_label/get-28    6.857n ± 0%   5.686n ± 0%  -17.08% (p=0.002 n=6)
Labels_Get/with_10_labels/not-found_label/has-28    6.909n ± 0%   5.570n ± 1%  -19.38% (p=0.002 n=6)
Labels_Get/with_30_labels/first_label/get-28        4.900n ± 1%   4.072n ± 0%  -16.91% (p=0.002 n=6)
Labels_Get/with_30_labels/first_label/has-28        3.351n ± 1%   2.983n ± 0%  -11.00% (p=0.002 n=6)
Labels_Get/with_30_labels/middle_label/get-28       40.76n ± 1%   33.17n ± 1%  -18.62% (p=0.002 n=6)
Labels_Get/with_30_labels/middle_label/has-28       40.65n ± 0%   31.83n ± 1%  -21.69% (p=0.002 n=6)
Labels_Get/with_30_labels/last_label/get-28         72.53n ± 1%   59.25n ± 0%  -18.32% (p=0.002 n=6)
Labels_Get/with_30_labels/last_label/has-28         73.41n ± 1%   57.15n ± 1%  -22.14% (p=0.002 n=6)
Labels_Get/with_30_labels/not-found_label/get-28    6.861n ± 1%   5.681n ± 1%  -17.20% (p=0.002 n=6)
Labels_Get/with_30_labels/not-found_label/has-28    6.928n ± 1%   5.575n ± 0%  -19.53% (p=0.002 n=6)
Labels_Equals/equal-28                              20.66n ± 1%   19.54n ± 1%   -5.42% (p=0.002 n=6)
Labels_Equals/not_equal-28                          17.38n ± 1%   16.43n ± 1%   -5.47% (p=0.002 n=6)
Labels_Equals/different_sizes-28                    11.53n ± 0%   10.12n ± 1%  -12.23% (p=0.002 n=6)
Labels_Equals/lots-28                               82.61n ± 0%   71.60n ± 0%  -13.33% (p=0.002 n=6)
Labels_Equals/real_long_equal-28                    76.01n ± 0%   73.55n ± 1%   -3.23% (p=0.002 n=6)
Labels_Equals/real_long_different_end-28            74.29n ± 0%   72.83n ± 1%   -1.97% (p=0.002 n=6)
Labels_Compare/equal-28                             20.00n ± 1%   19.48n ± 1%   -2.55% (p=0.002 n=6)
Labels_Compare/not_equal-28                         17.98n ± 1%   17.45n ± 1%   -2.95% (p=0.002 n=6)
Labels_Compare/different_sizes-28                  10.030n ± 1%   9.637n ± 0%   -3.92% (p=0.002 n=6)
Labels_Compare/lots-28                              92.99n ± 0%   78.31n ± 0%  -15.79% (p=0.002 n=6)
Labels_Compare/real_long_equal-28                   74.19n ± 0%   75.32n ± 1%   +1.52% (p=0.002 n=6)
Labels_Compare/real_long_different_end-28           71.47n ± 0%   72.34n ± 1%   +1.21% (p=0.002 n=6)
Labels_Hash/typical_labels_under_1KB-28             105.2n ± 1%   106.9n ± 1%   +1.66% (p=0.002 n=6)
Labels_Hash/bigger_labels_over_1KB-28               129.1n ± 0%   131.1n ± 1%   +1.59% (p=0.002 n=6)
Labels_Hash/extremely_large_label_value_10MB-28     491.7µ ± 1%   496.2µ ± 1%        ~ (p=0.065 n=6)
Labels_Copy-28                                      13.72n ± 1%   13.82n ± 1%   +0.73% (p=0.002 n=6)
geomean                                             22.88n        20.07n       -12.29%
```